# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Loop API",
    "version": "0.1.0",
    "description": "The Loop Autonomous Improvement Engine API. Collects signals, organizes work into issues, and dispatches AI agents with hydrated prompts.",
    "contact": {
      "name": "Loop by Dork Labs",
      "url": "https://www.looped.me"
    },
    "license": {
      "name": "MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5667",
      "description": "Local development"
    },
    {
      "url": "https://api.looped.me",
      "description": "Production"
    }
  ],
...
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "API key passed as Bearer token in the Authorization header"
      }
    },
    "schemas": {
...
      "PromptTemplate": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "slug": {
            "type": "string",
            "example": "signal-triage"
          },
          "name": {
            "type": "string",
            "example": "Signal Triage"
          },
          "description": {
            "type": ["string", "null"],
            "example": "Default template for triaging incoming signals."
          },
          "conditions": {
            "type": "object",
            "properties": {
              "type": {
                "type": "string",
                "enum": ["signal", "hypothesis", "plan", "task", "monitor"]
              },
              "signalSource": {
                "type": "string"
              },
              "labels": {
                "type": "array",
                "items": {
                  "type": "string"
                }
              },
              "projectId": {
                "type": "string"
              },
              "hasFailedSessions": {
                "type": "boolean"
              },
              "hypothesisConfidence": {
                "type": "number",
                "minimum": 0,
                "maximum": 1
              }
            },
            "description": "Conditions that determine when this template is selected",
            "example": {
              "type": "signal"
            }
          },
          "specificity": {
            "type": "integer",
            "minimum": 0,
            "maximum": 100,
            "example": 10
          },
          "projectId": {
            "type": ["string", "null"],
            "example": "cuid2_abc123"
          },
          "activeVersionId": {
            "type": ["string", "null"],
            "example": "cuid2_abc123"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          },
          "deletedAt": {
            "type": ["string", "null"],
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "slug",
          "name",
          "description",
          "conditions",
          "specificity",
          "projectId",
          "activeVersionId",
          "createdAt",
          "updatedAt",
          "deletedAt"
        ]
      },
      "PromptVersion": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "templateId": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "version": {
            "type": "integer",
            "example": 1
          },
          "content": {
            "type": "string",
            "example": "# Signal Triage: {{issue.title}}\n\nYou are triaging signal #{{issue.number}}."
          },
          "changelog": {
            "type": ["string", "null"],
            "example": "Initial version"
          },
          "authorType": {
            "type": "string",
            "enum": ["human", "agent"],
            "example": "human"
          },
          "authorName": {
            "type": "string",
            "example": "system"
          },
          "status": {
            "type": "string",
            "enum": ["active", "draft", "retired"],
            "example": "active"
          },
          "usageCount": {
            "type": "integer",
            "example": 0
          },
          "completionRate": {
            "type": ["number", "null"],
            "example": null
          },
          "avgDurationMs": {
            "type": ["number", "null"],
            "example": null
          },
          "reviewScore": {
            "type": ["number", "null"],
            "example": null
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "templateId",
          "version",
          "content",
          "changelog",
          "authorType",
          "authorName",
          "status",
          "usageCount",
          "completionRate",
          "avgDurationMs",
          "reviewScore",
          "createdAt"
        ]
      },
      "PromptReview": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "versionId": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "issueId": {
            "type": "string",
            "example": "cuid2_abc123"
          },
          "clarity": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 4
          },
          "completeness": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 5
          },
          "relevance": {
            "type": "integer",
            "minimum": 1,
            "maximum": 5,
            "example": 3
          },
          "feedback": {
            "type": ["string", "null"],
            "example": "Instructions were clear but missing error handling steps."
          },
          "authorType": {
            "type": "string",
            "enum": ["human", "agent"],
            "example": "agent"
          },
          "createdAt": {
            "type": "string",
            "format": "date-time",
            "example": "2026-02-20T12:00:00.000Z"
          }
        },
        "required": [
          "id",
          "versionId",
          "issueId",
          "clarity",
          "completeness",
          "relevance",
          "feedback",
          "authorType",
          "createdAt"
        ]
      }
    },
...
  },
  "paths": {
...
    "/api/templates/{id}/reviews": {
      "get": {
        "tags": ["Templates"],
        "summary": "List reviews for a template",
        "description": "Returns reviews across all versions of the template.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "schema": {
              "type": "string",
              "example": "cuid2_abc123"
            },
            "required": true,
            "name": "id",
            "in": "path"
          },
          {
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 200,
              "default": 50,
              "description": "Maximum records to return (1–200)",
              "example": 50
            },
            "required": false,
            "description": "Maximum records to return (1–200)",
            "name": "limit",
            "in": "query"
          },
          {
            "schema": {
              "type": ["integer", "null"],
              "minimum": 0,
              "default": 0,
              "description": "Number of records to skip",
              "example": 0
            },
            "required": false,
            "description": "Number of records to skip",
            "name": "offset",
            "in": "query"
          }
        ],
        "responses": {
          "200": {
            "description": "Paginated list of reviews",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/PromptReview"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "example": 42
                    }
                  },
                  "required": ["data", "total"]
                }
              }
            }
          },
          "404": {
            "description": "Template not found"
          }
        }
      }
    },
...
    "/api/prompt-reviews": {
      "post": {
        "tags": ["Prompt Reviews"],
        "summary": "Submit a prompt review",
        "description": "Submits a quality review for a prompt version. Updates the version's EWMA review score and triggers an improvement issue if quality degrades below threshold.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "versionId": {
                    "type": "string",
                    "example": "cuid2_abc123",
                    "description": "The prompt version being reviewed"
                  },
                  "issueId": {
                    "type": "string",
                    "example": "cuid2_abc123",
                    "description": "The issue this prompt was used for"
                  },
                  "clarity": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 4
                  },
                  "completeness": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 5
                  },
                  "relevance": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 5,
                    "example": 3
                  },
                  "feedback": {
                    "type": "string",
                    "example": "Instructions were clear but missing error handling steps."
                  },
                  "authorType": {
                    "type": "string",
                    "enum": ["human", "agent"],
                    "example": "agent"
                  }
                },
                "required": [
                  "versionId",
                  "issueId",
                  "clarity",
                  "completeness",
                  "relevance",
                  "authorType"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Review submitted",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "$ref": "#/components/schemas/PromptReview"
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "404": {
            "description": "Version not found"
          },
          "422": {
            "description": "Validation error"
          }
        }
      }
    },
...
    "/api/dashboard/prompts": {
      "get": {
        "tags": ["Dashboard"],
        "summary": "Template health with scores",
        "description": "Returns all prompt templates with version history and aggregated review scores.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Prompt health data",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "data": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "template": {
                            "$ref": "#/components/schemas/PromptTemplate"
                          },
                          "activeVersion": {
                            "allOf": [
                              {
                                "$ref": "#/components/schemas/PromptVersion"
                              },
                              {
                                "type": ["object", "null"]
                              }
                            ]
                          },
                          "recentVersions": {
                            "type": "array",
                            "items": {
                              "$ref": "#/components/schemas/PromptVersion"
                            },
                            "description": "Last 5 versions"
                          },
                          "reviewSummary": {
                            "type": "object",
                            "properties": {
                              "totalReviews": {
                                "type": "integer",
                                "example": 12
                              },
                              "avgClarity": {
                                "type": ["number", "null"],
                                "example": 4.1
                              },
                              "avgCompleteness": {
                                "type": ["number", "null"],
                                "example": 3.8
                              },
                              "avgRelevance": {
                                "type": ["number", "null"],
                                "example": 4.3
                              },
                              "compositeScore": {
                                "type": ["number", "null"],
                                "example": 4.07
                              }
                            },
                            "required": [
                              "totalReviews",
                              "avgClarity",
                              "avgCompleteness",
                              "avgRelevance",
                              "compositeScore"
                            ]
                          },
                          "needsAttention": {
                            "type": "boolean",
                            "description": "True when compositeScore < 3.0 or completionRate < 50%",
                            "example": false
                          }
                        },
                        "required": [
                          "template",
                          "activeVersion",
                          "recentVersions",
                          "reviewSummary",
                          "needsAttention"
                        ]
                      }
                    }
                  },
                  "required": ["data"]
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
...
  },
...
}